--- extra/mockgpu/amdgpu.py
@@ -1,38 +1,0 @@
-gpu7900xtx_props = """cpu_cores_count 0
-simd_count 192
-mem_banks_count 1
-caches_count 206
-io_links_count 1
-p2p_links_count 5
-cpu_core_id_base 0
-simd_id_base 2147488032
-max_waves_per_simd 16
-lds_size_in_kb 64
-gds_size_in_kb 0
-num_gws 64
-wave_front_size 32
-array_count 12
-simd_arrays_per_engine 2
-cu_per_simd_array 8
-simd_per_cu 2
-max_slots_scratch_cu 32
-gfx_target_version 110000
-vendor_id 4098
-device_id 29772
-location_id 34304
-domain 0
-drm_render_minor {drm_render_minor}
-hive_id 0
-num_sdma_engines 2
-num_sdma_xgmi_engines 0
-num_sdma_queues_per_engine 6
-num_cp_queues 8
-max_engine_clk_fcompute 2482
-local_mem_size 0
-fw_version 2140
-capability 671588992
-debug_prop 1495
-sdma_fw_version 20
-unique_id 11673270660693242239
-num_xcc 1
-max_engine_clk_ccompute 2400"""
